[UPDATE] Rename utils -> invoiceUtils
